Deduplicate the ConceptSchemeMultiSelectCheckboxes component
