+++ addon/components/rdf-input-fields/concept-scheme-multi-select-checkboxes.hbs
@@ -1,4 +1,9 @@
-<AuLabel for={{this.inputId}}>
+{{! TODO: Replace this with a fieldset + legend: https://github.com/appuniversum/ember-appuniversum/issues/210 }}
+<AuLabel
+  @error={{this.hasErrors}}
+  @required={{this.isRequired}}
+  for={{this.inputId}}
+>
   {{@field.label}}
 </AuLabel>
 
@@ -16,9 +21,14 @@
         @name={{@field.uri.value}}
         @label={{option.label}}
         @value={{option.label}}
-        @disabled={{true}}
+        @onChange={{fn this.updateValue option}}
+        @disabled={{@show}}
         checked={{option.provided}}
       />
     </div>
   {{/each}}
 </div>
+
+{{#each this.errors as |error|}}
+  <AuHelpText @error={{true}}>{{error.resultMessage}}</AuHelpText>
+{{/each}}
