--- addon/components/rdf-input-fields/concept-scheme-multi-select-checkboxes/edit.hbs
@@ -1,33 +1,0 @@
-{{! TODO: Replace this with a fieldset + legend: https://github.com/appuniversum/ember-appuniversum/issues/210 }}
-<AuLabel
-  @error={{this.hasErrors}}
-  @required={{this.isRequired}}
-  for={{this.inputId}}
->
-  {{@field.label}}
-</AuLabel>
-
-<div id={{this.inputId}} class="au-o-grid au-o-grid--flush">
-  {{#each this.options as |option|}}
-    <div
-      class="au-o-grid__item
-        {{if
-          this.isColumnLayout
-          'au-u-1-1@tiny au-u-1-2@small au-u-1-3@medium'
-        }}"
-    >
-      <AuControlCheckbox
-        @identifier="{{@field.uri}}/<{{option.subject.value}}"
-        @name={{@field.uri.value}}
-        @label={{option.label}}
-        @value={{option.label}}
-        @onChange={{fn this.updateValue option}}
-        checked={{option.provided}}
-      />
-    </div>
-  {{/each}}
-</div>
-
-{{#each this.errors as |error|}}
-  <AuHelpText @error={{true}}>{{error.resultMessage}}</AuHelpText>
-{{/each}}
