+++ app/components/rdf-input-fields/concept-scheme-multi-select-checkboxes.js
@@ -1,0 +1,1 @@
+export { default } from '@lblod/ember-submission-form-fields/components/rdf-input-fields/concept-scheme-multi-select-checkboxes';
