--- app/components/rdf-input-fields/concept-scheme-multi-select-checkboxes/show.js
@@ -1,1 +1,0 @@
-export { default } from '@lblod/ember-submission-form-fields/components/rdf-input-fields/concept-scheme-multi-select-checkboxes/show';
